client/requirements.txt
colorama==0.4.6
pillow==10.4.0
pynput==1.7.7
python_socketio==5.11.4
requests==2.32.2